Fix input type for monthly income and expenses to float for accurate calculations

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-user_monthly_income = float(input("Enter your monthly income:"))
-user_monthly_expense = float(input("Enter your total monthly expenses:"))
+user_monthly_income = int(input("Enter your monthly income:"))
+user_monthly_expense = int(input("Enter your total monthly expenses:"))
 
 monthly_savings = user_monthly_income - user_monthly_expense
 interest_rate = 0.05
